[feature/post-paging]: Login Dialog 개발

+++ common-client/src/commonMain/kotlin/commonClient/domain/entity/post/Post.kt
@@ -8,7 +8,7 @@ data class Post(
     val author: User,
     val voteResponses: List<Vote>,
     val categories: List<Category>,
-    val selectedVote: Int?,
+    val selectedVote: Long?,
     val isLiked: Boolean?,
     val createdAt: String,
     val updatedAt: String?
